requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
PyMuPDF==1.24.0
pytesseract==0.3.10
sentence-transformers==2.7.0
faiss-cpu==1.8.0
langchain==0.1.16
python-dotenv==1.0.1
transformers==4.41.0
accelerate==0.29.3
pdfplumber==0.10.4
Pillow==10.3.0